// src/common/enums/AppUser/UserRole.ts
export enum UserRole {
  Admin = 1,
  WarehouseManager = 2,
  WarehouseStaff = 3,
  SalesStaff = 4,
  DeliveryStaff = 5,
  Accountant = 6,
  Auditor = 7
}

export const CustomRoles = {
  Admin: "Admin",
  WarehouseManager: "WarehouseManager",
  WarehouseStaff: "WarehouseStaff",
  SalesStaff: "SalesStaff",
  DeliveryStaff: "DeliveryStaff",
  Accountant: "Accountant",
  Auditor: "Auditor"
};